+++ vscode-dotnet-runtime-library/src/Acquisition/DotnetResolver.ts
@@ -3,12 +3,16 @@
 *  Licensed to the .NET Foundation under one or more agreements.
 *  The .NET Foundation licenses this file to you under the MIT license.
 *--------------------------------------------------------------------------------------------*/
-import { DotnetUnableToCheckPATHArchitecture } from '../EventStream/EventStreamEvents';
+import * as os from 'os';
+import * as path from 'path';
+import { DotnetFindPathNoRuntimesOnHost, DotnetUnableToCheckPATHArchitecture } from '../EventStream/EventStreamEvents';
+import { LocalMemoryCacheSingleton } from '../LocalMemoryCacheSingleton';
 import { CommandExecutor } from '../Utils/CommandExecutor';
 import { CommandExecutorCommand } from '../Utils/CommandExecutorCommand';
 import { ExecutableArchitectureDetector } from '../Utils/ExecutableArchitectureDetector';
 import { ICommandExecutor } from '../Utils/ICommandExecutor';
 import { IUtilityContext } from '../Utils/IUtilityContext';
+import { getDotnetExecutable } from '../Utils/TypescriptUtilities';
 import { DOTNET_INFORMATION_CACHE_DURATION_MS } from './CacheTimeConstants';
 import { DotnetCoreAcquisitionWorker } from './DotnetCoreAcquisitionWorker';
 import { DotnetInstallMode } from './DotnetInstallMode';
@@ -32,8 +36,6 @@ export class DotnetResolver implements IDotnetResolver
      */
     public async getDotnetInstalls(dotnetExecutablePath: string, mode: DotnetInstallMode, requestedArchitecture: string | undefined | null): Promise<IDotnetListInfo[]>
     {
-        // TODO: Bug 1 - Does not resolve path such as 'dotnet' below
-        const hostArch = new ExecutableArchitectureDetector().getExecutableArchitecture(dotnetExecutablePath);
         const oldLookup = process.env.DOTNET_MULTILEVEL_LOOKUP;
         // This is deprecated but still needed to scan .NET 6 and below
         process.env.DOTNET_MULTILEVEL_LOOKUP = '0'; // make it so --list-runtimes only finds the runtimes on that path: https://learn.microsoft.com/en-us/dotnet/core/compatibility/deployment/7.0/multilevel-lookup#reason-for-change
@@ -42,9 +44,8 @@ export class DotnetResolver implements IDotnetResolver
         {
             if (mode === 'sdk')
             {
-                // BUG 2 - Somehow --arch invalid check is not running if hostArch is null
-                const availableSDKs = await this.getSDKs(dotnetExecutablePath, requestedArchitecture ?? DotnetCoreAcquisitionWorker.defaultArchitecture(), ExecutableArchitectureDetector.IsKnownArchitecture(hostArch));
-                const resolvedHostArchitecture = ExecutableArchitectureDetector.IsKnownArchitecture(hostArch) ? hostArch : availableSDKs?.at(0)?.architecture ?? await this.getHostArchitectureViaInfo(dotnetExecutablePath, requestedArchitecture);
+                const availableSDKs = await this.getSDKs(dotnetExecutablePath, requestedArchitecture ?? DotnetCoreAcquisitionWorker.defaultArchitecture());
+                const resolvedHostArchitecture = await this.resolveHostArchitecture(dotnetExecutablePath, availableSDKs, requestedArchitecture);
 
                 if (requestedArchitecture && (resolvedHostArchitecture !== requestedArchitecture))
                 {
@@ -63,8 +64,8 @@ export class DotnetResolver implements IDotnetResolver
             else
             {
                 // No need to consider SDKs when looking for runtimes as all the runtimes installed with the SDKs will be included in the runtimes list.
-                const availableRuntimes = await this.getRuntimes(dotnetExecutablePath, requestedArchitecture ?? DotnetCoreAcquisitionWorker.defaultArchitecture(), ExecutableArchitectureDetector.IsKnownArchitecture(hostArch));
-                const resolvedHostArchitecture = ExecutableArchitectureDetector.IsKnownArchitecture(hostArch) ? hostArch : availableRuntimes?.at(0)?.architecture ?? await this.getHostArchitectureViaInfo(dotnetExecutablePath, requestedArchitecture);
+                const availableRuntimes = await this.getRuntimes(dotnetExecutablePath, requestedArchitecture ?? DotnetCoreAcquisitionWorker.defaultArchitecture());
+                const resolvedHostArchitecture = await this.resolveHostArchitecture(dotnetExecutablePath, availableRuntimes, requestedArchitecture);
 
                 if (requestedArchitecture && (resolvedHostArchitecture !== requestedArchitecture))
                 {
@@ -79,6 +80,117 @@ export class DotnetResolver implements IDotnetResolver
                     };
                 }) as IDotnetListInfo[];
             }
+        }
+        finally
+        {
+            // Restore the environment variable to its original value
+            if (oldLookup !== undefined)
+            {
+                process.env.DOTNET_MULTILEVEL_LOOKUP = oldLookup;
+            }
+            else
+            {
+                delete process.env.DOTNET_MULTILEVEL_LOOKUP;
+            }
+        }
+    }
+
+    /**
+     * Determines the architecture of the .NET installation.
+     * @param dotnetExecutablePath The path to the dotnet executable - can be shell relative.
+     * @param availableInstalls The list of available .NET installations to the executable.
+     * @param requestedArchitecture The architecture requested by the user.
+     * @returns The determined architecture or null if it could not be determined.
+     */
+    private async resolveHostArchitecture(dotnetExecutablePath: string, availableInstalls: IDotnetListInfo[], requestedArchitecture: string | null | undefined): Promise<string | null>
+    {
+        // Resolve the true path, so a path such as 'dotnet' can be resolved as a file to get the architecture of, without needing to call 'which' or 'where'
+        const truePath = await this.resolveTruePath(dotnetExecutablePath, requestedArchitecture ?? null);
+
+        // Try to get the arch in the most performant way
+        const hostArch = new ExecutableArchitectureDetector().getExecutableArchitecture(truePath ?? dotnetExecutablePath);
+        const knownArchitecture = ExecutableArchitectureDetector.IsKnownArchitecture(hostArch);
+
+        if (knownArchitecture)
+        {
+            return hostArch;
+        }
+        else if (await this.hostSupportsArchFlag(truePath ?? dotnetExecutablePath, availableInstalls))
+        {
+            return requestedArchitecture ?? null;
+        }
+        else
+        {
+            return this.getHostArchitectureViaInfo(dotnetExecutablePath, requestedArchitecture);
+        }
+    }
+
+    /**
+     *
+     * @param tentativePath The path to resolve.
+     * @param suggestedArchitecture The architecture to resolve the paths for - this is only a suggestion and you must validate afterward
+     * @returns The resolved true path or null if not found.
+     */
+    public async resolveTruePath(tentativePath: string, suggestedArchitecture: string | null): Promise<string | null>
+    {
+        const truePaths = await this.resolveTruePaths([tentativePath], suggestedArchitecture);
+        return truePaths?.[0] ?? null;
+    }
+
+    /**
+    *
+    * @param tentativePaths Paths that may hold a dotnet executable.
+    * @param suggestedArchitecture The architecture to resolve the paths for - this is only a suggestion and you must validate afterward
+    * @returns The actual physical location/path on disk where the executables lie for each of the paths.
+    * Some of the symlinks etc resolve to a path which works but is still not the actual path.
+    */
+    public async resolveTruePaths(tentativePaths: string[], suggestedArchitecture: string | null): Promise<string[]>
+    {
+        const oldLookup = process.env.DOTNET_MULTILEVEL_LOOKUP;
+        // This is deprecated but still needed to scan .NET 6 and below
+        process.env.DOTNET_MULTILEVEL_LOOKUP = '0'; // make it so --list-runtimes only finds the runtimes on that path: https://learn.microsoft.com/en-us/dotnet/core/compatibility/deployment/7.0/multilevel-lookup#reason-for-change
+
+        try
+        {
+            const truePaths = [];
+            suggestedArchitecture ??= DotnetCoreAcquisitionWorker.defaultArchitecture()
+
+            for (const tentativePath of tentativePaths)
+            {
+                // This will even work if only the sdk is installed, list-runtimes on an sdk installed host would work
+                const runtimeInfo = await this.getRuntimes(tentativePath, suggestedArchitecture);
+                if ((runtimeInfo?.length ?? 0) > 0)
+                {
+                    // The .NET install layout is a well known structure on all platforms.
+                    // See https://github.com/dotnet/designs/blob/main/accepted/2020/install-locations.md#net-core-install-layout
+                    //
+                    // Therefore we know that the runtime path is always in <install root>/shared/<runtime name>
+                    // and the dotnet executable is always at <install root>/dotnet(.exe).
+                    //
+                    // Since dotnet --list-runtimes will always use the real assembly path to output the runtime folder (no symlinks!)
+                    // we know the dotnet executable will be two folders up in the install root.
+                    const truePath = path.join(path.dirname(path.dirname(runtimeInfo[0].directory)), getDotnetExecutable());
+                    truePaths.push(truePath);
+
+                    // Preload the cache with the calls we've already done.
+                    // Example: 'dotnet' --list-runtimes will be the same as 'C:\\Program Files\\dotnet\\dotnet.exe' --list-runtimes
+                    // If the dotnet executable full path was 'C:\\Program Files\\dotnet\\dotnet.exe'.
+
+                    // We do NOT want to do this on Unix, because the dotnet executable is potentially polymorphic.
+                    // /usr/local/bin/dotnet becomes /snap/dotnet-sdk/current/dotnet in reality, may have different behavior in shells.
+                    if (os.platform() === 'win32')
+                    {
+                        LocalMemoryCacheSingleton.getInstance().aliasCommandAsAnotherCommandRoot(`"${truePath}"`, `"${tentativePath}"`, this.workerContext.eventStream);
+                    }
+                }
+                else
+                {
+                    this.workerContext.eventStream.post(new DotnetFindPathNoRuntimesOnHost(`The host: ${tentativePath} does not contain a .NET runtime installation.`));
+
+                }
+            }
+
+            return (truePaths?.length ?? 0) > 0 ? truePaths : tentativePaths;
         }
         finally
         {
@@ -147,11 +259,10 @@ Please set the PATH to a dotnet host that matches the architecture. An incorrect
     /**
      *
      * @param existingPath the path to the executable of the dotnet muxer
-     * @param requestedArchitecture the architecture we want SDKs of
-     * @param knownArchitecture whether we already know the architecture of the host and therefore the SDKs - not a safe check if true is used
-     * @returns An array of IDotnetListInfo objects representing the SDKs installed on the system.
+     * @param requestedArchitecture the architecture we want SDKs of - not a strict requirement, you must validate afterward
+     * @returns An array of IDotnetListInfo objects representing the SDKs installed on the system - arch is always null
      */
-    public async getSDKs(existingPath: string, requestedArchitecture: string, knownArchitecture: boolean): Promise<IDotnetListInfo[]>
+    private async getSDKs(existingPath: string, requestedArchitecture: string): Promise<IDotnetListInfo[]>
     {
         if (!existingPath || existingPath === '""')
         {
@@ -167,7 +278,6 @@ Please set the PATH to a dotnet host that matches the architecture. An incorrect
             return [];
         }
 
-        const architectureKnown = knownArchitecture ? true : await this.hostSupportsArchFlag(existingPath, result.stdout);
         const sdks = result.stdout.split('\n').map((line) => line.trim()).filter(Boolean);
         const sdkInfos: IDotnetListInfo[] = sdks.map((sdk) =>
         {
@@ -176,14 +286,14 @@ Please set the PATH to a dotnet host that matches the architecture. An incorrect
                 mode: 'sdk',
                 version: parts[0],
                 directory: sdk.split(' ').slice(1).join(' ').slice(1, -1), // need to remove the brackets from the path [path],
-                architecture: architectureKnown ? requestedArchitecture : null
+                architecture: null
             } as IDotnetListInfo;
         }).filter(x => x !== null) as IDotnetListInfo[];
 
         return sdkInfos;
     }
 
-    private async hostSupportsArchFlag(dotnetExecutablePath: string, listDotnetInstallsStdout: string): Promise<boolean>
+    private async hostSupportsArchFlag(dotnetExecutablePath: string, availableInstalls: IDotnetListInfo[]): Promise<boolean>
     {
         // https://github.com/dotnet/runtime/pull/116078 --arch was not added until .NET 10 to allow us to skip calling dotnet --info because that is slow, as it is not native code.
         // However, --arch gets ignored if the host does not support it. The output is also identical with or without --arch.
@@ -192,7 +302,7 @@ Please set the PATH to a dotnet host that matches the architecture. An incorrect
         // The --arch flag was added in the middle of .NET 10, so we can assume it is supported if the version is 10.0 or later.
         // We don't want to slow down the current common case for people without .NET 10 by adding another process spawn check.
         // We don't check that the version is 10.0 or later after 2026 when .NET 11 starts rolling out, as It will be slower to check all of the numbers in the output for versions >= 10.
-        const hostMaySupportArchFlag = listDotnetInstallsStdout.includes("10.0") || listDotnetInstallsStdout.includes("11.0") || Date.now() >= new Date('2026-03-01').getTime();
+        const hostMaySupportArchFlag = availableInstalls.some(install => install.version.startsWith("10.0") || install.version.startsWith("11.0")) || Date.now() >= new Date('2026-03-01').getTime();
         // Use runtimes instead of sdks, as sdks will always have a runtime, and runtime search can be cached across both mode calls.
         const findInvalidCommand = CommandExecutor.makeCommand(`"${dotnetExecutablePath}"`, ['--list-runtimes', '--arch', 'invalid-arch']);
         const hostSupportsArchFlag = hostMaySupportArchFlag ? (await (this.executor!).execute(findInvalidCommand, { dotnetInstallToolCacheTtlMs: DOTNET_INFORMATION_CACHE_DURATION_MS }, false)).status !== '0' : false;
@@ -207,11 +317,10 @@ Please set the PATH to a dotnet host that matches the architecture. An incorrect
     /**
      *
      * @param existingPath the path to the executable of the dotnet muxer
-     * @param requestedArchitecture the architecture we want runtimes of - not a safe check if true is used
-     * @param knownArchitecture whether the architecture is known
-     * @returns an array of IDotnetListInfo objects representing the runtimes installed on the system
+     * @param requestedArchitecture the architecture we want runtimes of - not a strict requirement, you must validate afterward
+     * @returns an array of IDotnetListInfo objects representing the runtimes installed on the system - arch is always null
      */
-    public async getRuntimes(existingPath: string, requestedArchitecture: string | null, knownArchitecture: boolean): Promise<IDotnetListInfo[]>
+    public async getRuntimes(existingPath: string, requestedArchitecture: string | null): Promise<IDotnetListInfo[]>
     {
         if (!existingPath || existingPath === '""')
         {
@@ -231,7 +340,6 @@ Please set the PATH to a dotnet host that matches the architecture. An incorrect
             return [];
         }
 
-        const architectureKnown = knownArchitecture ? true : await this.hostSupportsArchFlag(existingPath, result.stdout);
         const runtimes = result.stdout.split('\n').map((line) => line.trim()).filter(Boolean);
         const runtimeInfos: IDotnetListInfo[] = runtimes.map((runtime) =>
         {
@@ -241,7 +349,7 @@ Please set the PATH to a dotnet host that matches the architecture. An incorrect
                 version: parts[1],
                 directory: runtime.split(' ').slice(2).join(' ').slice(1, -1), // account for spaces in PATH, no space should appear before then and luckily path is last.
                 // the 2nd slice needs to remove the brackets from the path [path]
-                architecture: architectureKnown ? requestedArchitecture : null
+                architecture: null
             } as IDotnetListInfo;
         }).filter(x => x !== null) as IDotnetListInfo[];
