+++ vscode-dotnet-runtime-library/src/index.ts
@@ -8,9 +8,9 @@ export * from './Acquisition/DirectoryProviderFactory';
 export * from './Acquisition/DotnetConditionValidator';
 export * from './Acquisition/DotnetCoreAcquisitionWorker';
 export * from './Acquisition/DotnetCoreDependencyInstaller';
+export * from './Acquisition/DotnetHostPathFinder';
 export * from './Acquisition/DotnetInstall';
 export * from './Acquisition/DotnetInstallMode';
-export * from './Acquisition/DotnetPathFinder';
 export * from './Acquisition/DotnetResolver';
 export * from './Acquisition/ExistingPathResolver';
 export * from './Acquisition/GenericDistroSDKProvider';
